Moved channel handler to module
